requirements.txt
psycopg2-binary
python-dotenv
basedosdados
google-cloud-bigquery
pydata-google-auth
